Move test_lib under app/lib/
Did this because the IDE is not smart enough to recognize test_lib as a separate library
Also, the relative imports were annoying to deal with
