+++ app/lib/core/services/database/database_service.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:app/core/models/animal_transport_record.dart';
-import 'package:app/core/models/test_animal_transport_record.dart';
+import 'package:app/test/test_animal_transport_record.dart';
 
 class DatabaseService {
   // TODO: #130. Replace with actual DB functions
