+++ app/test/ui/views/active/form_field/shipper_info_form_field_test.dart
@@ -1,10 +1,9 @@
 import 'package:app/core/models/address.dart';
 import 'package:app/core/models/shipper_info.dart';
+import 'package:app/test/test_animal_transport_record.dart';
 import 'package:app/ui/views/active/form_field/shipper_info_form_field.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import '../../../../test_lib/test_animal_transport_record.dart';
 
 void main() {
   void verifyInformationIsShown(ShipperInfo infoExpected) {
