+++ app/test/ui/widgets/atr_preview_card_test.dart
@@ -1,9 +1,8 @@
+import 'package:app/test/test_animal_transport_record.dart';
 import 'package:app/ui/widgets/atr_preview_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:intl/intl.dart';
-
-import '../../test_lib/test_animal_transport_record.dart';
 
 void main() {
   group('ATR Preview Card', () {
